pyproject.toml
[project]
name = "decomp"
version = "0.1.0"
description = "N64 decompilation agent — ROM-agnostic tool using Claude for iterative decompilation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "asm-differ @ git+https://github.com/simonlindholm/asm-differ.git",
    "m2c @ git+https://github.com/matt-kempster/m2c.git",
    "splat64[mips]>=0.39.0,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
decomp = "decomp.main:main"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-first-party = ["decomp"]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
